Make class SQLSHELL accessible outside of the mssqlclient.py example (#1535)
* Make class SQLSHELL accessible outside of the mssqlclient.py example

* Separate mssqlshell.py from mssqlclient.py

* Fix import

* Add interactive SQL shell to MSSQL attack in ntlmrelayx
